Explicitly pick C-b as tmux prefix
When a community maintained Chef cookbook was installed on my dev box,
it set up a system-wide tmux.conf (/etc/tmux.conf) that unbound the
default prefix (C-b) to set up a prefix key that matches Screen (C-a).
Since I've been using tmux, I have come to enjoy the C-b prefix, as it
frees up C-a for use as Readline's "go to start of line" command.

Explicitly selecting C-b should do nothing on the stock install, and
will override the global binding when someone has rudely changed it.