Use `push-input` instead of commenting zsh lines
I have a habit of writing part of a zsh line, and then wanting to save
it, run something else, and then come back to it. In Bash, I would
comment out the line, run the line, run my other command, and then go up
in history to find my original comment, remove the comment prefix, and
run the command.

That's somewhat roundabout; here's a better answer. Now I press ^x, the
line I am editing is pushed onto a stack and replaced with a blank
prompt. After I run a command, the top result from the input stack is
pushed back. This is a much nicer experience.