Open tmux panes to the current pane's directory
It's common for me to have different windows in different working
directories. The usual case is that when I perform a split, I want to
have another pane in the same directory I'm already in (for example, to
display a file in this directory that I'm working in).

Set my split / new window commands to have this behavior.