Add work identity to gitconfig
Previously, I've been using an override in `.gitconfig.local` on my work
machine, but that meant I still had to be careful if I was ever coding
on my personal laptop to set up my git user info to the right identity
on a per-repo basis. As of git 2.13, conditional includes give me a
better solution for this (especially since I already keep all my work
stuff in a specific subdirectory anyways).