Add Hustle Kafka helper script to get consumer group info
This command is a pain to type out, and even more of a pain to use well
because you have to get the consumer group name from another obscure
invocation of `kafka-consumer-groups.sh`. By wrapping it in a little
script, I can also add a ZSH autocompletion to make this operation easy
for me to do.